Inline KvStoreCombinedDaoCommon.java
